requirements.txt
streamlit
matplotlib
pandas
requests
Pillow
google-generativeai
